Move parseGeometry and parseGeometryPoint tests to geo
